ash: standardise treatment of winxp option
Although 'winxp' is a shell option it could only be set with '-X'
on the command line.

Fully implement 'winxp' so it can also be set within the shell by
'set -o winxp' and 'set +o winxp'.  '-X' no longer needs to be the
first option on the command line.

Track which shell variables have been imported from a native Windows
environment so only those are affected when 'winxp' is changed.  The
tracking persists in a subshell but is lost when shell variables are
exported to the environment so 'set -/+o winxp' is ineffective in a
child shell.

Costs 48-52 bytes.

(GitHub issue #322)