win32: make support for euro input a separate option
Commit 93a63809f (win32: add support for the euro currency symbol)
made various changes to enable support for the euro symbol.

One of these changes allows the euro to be entered from the console
even if the current code page doesn't support it.  This is probably
of limited use:  the symbol can be entered but won't be displayed
correctly.

Move this capability into a separate configuration option,
FEATURE_EURO_INPUT, which is disabled by default.

Saves 48-64 bytes in the new default case.

(GitHub issue #335)