win32: avoid crashing the console with poll(2)
Commit 8e6991733 (ash: fix 'read' shell built-in (1)) introduced
the use of poll(2) in the shell 'read' built-in.  When the UTF8
code page is in use this results in the console crashing if a 3 or
more byte UTF8 character is entered.

The crash is caused by the use of PeekConsoleInputA() which, like
ReadConsoleInputA(), is broken.  It can be avoided by using
PeekConsoleInputW() instead.

The number of key events will differ but this doesn't matter in
this case as poll(2) effectively runs in a busy loop with a 1ms
sleep.