win32: UTF8 input: improve missing-key-down hack
The UTF8 input code works around an issue when pasting at the
windows console (but not terminal) that sometimes we get key-up
without a prior matching key-down - at which case it generates down.

However, previously it detected this by comparing an up-event to the
last down-event, which could result in false-positive in cases like:
  X-down Y-down X-up Y-up (e.g. when typing quickly).

Now it remembers the last 8 key-down events when searching a prior
matching key-down, which fixes an issue of incorrect repeated keys
(in the example above Y-up was incorrectly changed to Y-down).