win32: fail early if UTF8 isn't supported
When the UTF8 manifest is included in the binary and ACP isn't UTF8
fail at once.  This avoids raising false hopes if the binary is run
on Window 7/8.  On Windows XP it won't even run.

When the busybox applet is run without arguments always report the
build-time configuration of globbing and the UTF8 manifest.