diff: more changes to --binary
The changes introduced to support the --binary option gave incorrect
results when comparing files with CRLF line endings *without* the
--binary option present.

The code needs to keep track of the position within the file and is
confused by text mode.

As an alternative solution, always use binary mode but skip the CR
of a CRLF pair when the --binary option isn't used.  This gives
results matching GNU diff when comparing files with matching line
endings, with or without --binary.  When line endings differ the
results aren't always the same.

Costs 32 bytes in the 32-bit build, saves 16 in 64-bit.

(GitHub issue #348)