win32: replace readlink(2)
The Windows implementation of readlink(2) has caused problems in
the past.  As, for example, with commit c29dc205d2 (win32: fix
implementation of readlink(2)).

Most uses of readlink(2) in BusyBox are actually calls to the
(considerably more convenient) library function xmalloc_readlink().
Implement a Windows version of that and used it instead of readlink(2).

This improves the handling of symbolic links (and similar reparse
points) in CJK and UTF-8 code pages.

Saves 48-80 bytes.