make: allow targets of the form c:/path
Parsing of target rules was too simplistic to allow targets of
the form c:/path.

Allow such targets as a non-POSIX extension.  Also add a 'windows'
pragma to allow such targets in POSIX mode.  This additionally
requires the use of the 'target_name' pragma to permit slashes in
target names.

This change *doesn't* allow targets of the form c:path, but it's
probably not wise to use them anyway.

Costs 40-80 bytes.

(GitHub issue #353)