make: fixes to warning messages
According to POSIX only fatal diagnostic messages should be written
to stderr.  Arrange for warning() to write to stdout.

pdpmake was rather too prolific in generating 'nothing to be done
for XXX' messages.  GNU make only issues the equivalent message
for top-level targets and only if nothing at all was (or would have
been) done.  bmake doesn't seem to have such a message.

Follow the practice of GNU make.  This requires more intensive
tracking of the actions taken while the make is in progress.

Costs 16 bytes.

(GitHub issue #354)