libbb: fix #!/bin/busybox
Commit 3476fb9f4c (ps: display interpreted script name in comm
column) required the applet start up code to detect the index
of the script in a shebang command line.

When a script starts with '#!/bin/busybox' the argument list is
shifted in busybox_main() but the script index wasn't adjusted
to match.  This caused the command to fail.

Make the proper adjustment.

Add 16 bytes.

(GitHub issue #357)