ps: increase length of 'comm' field
In Linux the command name associated with a process (as can be
obtained from '/proc/<PID>/comm') is truncated to 16 characters.
Thus upstream BusyBox only allows 16 characters for the 'comm'
field in 'ps'.

There's no need for such a constraint in busybox-w32.  Moreover,
the command name is used for the full command line ('args' field)
in most cases.  This field is allowed to be rather long in 'ps'
so it's not expected to be truncated.

Still, to avoid diverging too much from upstream it's best to have
some measure of truncation.  Increase the allowed length of the
command name to 32 characters.

Adds 16 bytes.

(GitHub issue #358)