ash: changes to login shell functionality
Upstream BusyBox supports the usual Unix functionality where the
login process adds a '-' prefix to the shell's argv[0].  Since
busybox-w32 doesn't have a login process this serves no purpose
and can be removed.

Instead it would be useful to have a way to invoke a login shell
without requiring any arguments.  Advantages would be:

- easier to configure busybox-w32 'ash' as the SSH login shell;

- double clicking on a shell binary could start a login shell.

Add 'lash' as an alias for 'ash' which starts a login shell without
requiring the '-l' option.

Saves 16 bytes.