make: avoid more warnings
The tracking of progress introduced in commit fb4be267c (make: fixes
to warning messages) failed to account for runs with the '-t' or '-n'
options.  Without this the following makefile:

   dummy_target: dummy_rule

   dummy_rule:
       @:

issues the warning 'make: nothing to be done for dummy_target' when
run with '-t' or '-n' and no 'dummy_rule' file present.

(GitHub issue #354)