ash: fix signal handling
Commit 89aa9d783 (win32: changes to signal handling) caused a
regression:  'httpd -f' could no longer be interrupted by Ctrl-C.

The specific issue is the use of exit(2) to replace raise(3) in
raise_interrupt().  This change was made so the shell would return
the expected exit code of 130 rather than the default code of 3
which Windows uses when a signal with SIG_DFL disposition is caught.

The problem can be avoided by calling _exit(2) instead.  A call to
fflush_all() has also been added, as _exit(2) doesn't do that.

(GitHub issue #365)