make: permit Unix-style paths when setting MAKE
The 'make' applet follows POSIX requirements when setting the MAKE
variable.  However, it doesn't allow for the case where argv[0] is
of a form like '/bin/make' but no corresponding executable exists.
This can happen in busybox-w32 when '/bin/make' is interpreted as
a reference to the 'make' applet.

In this case set the MAKE variable to argv[0] and avoid issuing a
warning.

Setting MAKE to something that isn't a real executable is fine so
long as it's only used by busybox-w32 applets.  If it's used by
external applications they may get confused.

Adds 16-32 bytes.

(GitHub issue #354)