win32: avoid terminal weirdness induced by Gradle?
GitHub issue #372 reports that in certain circumstances (which I've
been unable to reproduce) Gradle leaves the terminal in a state
where linefeeds seem not to result in a carriage return.

This *might* be because Gradle sets DISABLE_NEWLINE_AUTO_RETURN in
the terminal mode.  Reset DISABLE_NEWLINE_AUTO_RETURN to zero before
the shell prompt is issued to see of this makes any difference.

Costs 16-32 bytes.