make: fix detection of target rules
The presence of an equal sign in an inline command on a target
rule caused the line to be detected as a macro assignment.

For example:

   target:; @echo a = $(a)

Rearrange input parsing so target rules are detected before macro
assignments.  This is made more complex by having to allow for the
':=', '::=' and ':::=' assignment operators.  (And for targets
containing colons on Windows.)

Costs 240-248 bytes.