lineedit: normalise HOME for use in prompt
The code to replace the user's home directory with '~' in the
prompt didn't allow for the possibility of the HOME environment
variable having backslashes in the path.

Convert backslashes to forward slashes in HOME.  This isn't necessary
if the home directory is obtained from getpwuid(3) as it's already
converted there.

Adds 24-32 bytes.