win32: only search PATH for compressor
mingw_fork_compressor() uses CreateProcess() to run the compressor
program.  This will often be an instance of BusyBox, but since the
xv and lzma applets in BusyBox don't support compression it can be
an external program.

It was intended that the external program should be found using PATH.
However, CreateProcess() looks in various other places before trying
PATH.  In particular, it first looks in the directory of the current
executable, then in the current directory of the process.  This can
result in the wrong xz.exe or lzma.exe being found.

Perform an explicit PATH search and force CreateProcess() to use the
result.

This change only affects the search for a compressor.  The same
problem also affects other uses of our popen(3) emulation.  These
may be addressed in future.

Costs 64-80 bytes.

(GitHub issue #376)