win32: allow hardcoded list of applets to override
The environment variable BB_OVERRIDE_APPLETS provides control over
which applets are preferred in standalone shell mode.  Allow a
similar list to be hardcoded in the binary with the OVERRIDE_APPLETS
configuration string.

The environment variable is checked first.  If it doesn't override
the applet the hardcoded string is then checked too.

The default for OVERRIDE_APPLETS is an empty string.  In this case
the size and behaviour of the binary is unchanged.