lineedit: fix to tab completion of applet override
When the ability to override an applet if an external binary was
present was added, the commit message for bd7018350 said:

   This doesn't affect tab completion in the shell:  whether a
   completion is an applet or an external command is irrelevant.

This isn't quite true.  Suppose 'applet' has been overridden in
such circumstances:

   $ export BB_OVERRIDE_APPLETS=";applet"
   $ applet

If the user now enters tab twice the output would be:

   applet     applet.exe

The applet is still treated as a potential match.  Recent changes
to the applet override code result in a simple fix:  the shell PATH
can be passed to is_applet_preferred() in the line editing code.
Now only 'applet.exe' is treated as a match.

Adds 16-32 bytes.