httpd: fix failure to run daemonised
Since commit 2af141a2c (ash: changes to login shell functionality)
httpd has failed to run in the background (i.e. without the '-f'
option).

This was due the removal of support for applet names with a '-'
prefix.  Although this feature isn't used by the shell it was used
by httpd to mark when the process had been daemonised.

Adds 16 bytes.

(GitHub issue #378)