httpd: fix return code when run in background
When httpd was run in the background the return code of the parent
process was incorrect.  It seems when spawn() is run in _P_DETACH
mode it returns 0 on success, not a process handle.

Fix the test for the return code and alter mingw_spawn_detach()
so it doesn't treat the return from spawn() as a handle.

Saves 32 bytes.