httpd: enable support for CGI
The upstream code uses fork/exec when running a CGI process.
Emulate this by:

- Spawning a child httpd process with the special '-I 0' option,
  along with the options provided on the server command line.  This
  sets up the proper state then calls the cgi_handler() function.

- The cgi_handler() function fixes the pipe file descriptors and
  starts another child process to run the CGI script.

These processes are detached from the console on creation.  When
spawn() functions are run in P_DETACH mode they don't connect to
the standard file descriptors.  Normally this doesn't matter but
the process which runs the CGI scripts needs to inherit the pipe
endpoints.  The create_detached_process() function handles this.
See:

  https://github.com/rprichard/win32-console-docs/blob/master/README.md

Adds about 2.9Kb to the size of the binary.

(GitHub issue #266)