httpd: consistently leak memory, or not
create_detached_process() is only used when running a CGI script.
Previously it leaked the return values from quote_arg() but freed
the command line it built.

Whether or not the CGI script is successfully run its parent process
exits almost immediately, so there's no pressing need to free the
memory.  If FEATURE_CLEAN_UP is disabled (which it is by default)
don't bother.

Saves 16 bytes.