win32: make the clang build less crashy
busybox-w32 binaries built using clang crashed so frequently that
they were pretty much unusable.  The main issue seems to be with
assignments to the structures containing global variables which
are used in most applets.

Upstream commit 5156b2455 (Make const ptr assign as function call
in clang) addresses this, but is insufficient for the build on
Windows.  Extend the idea to the ASSIGN_CONST_PTR() macro too.

Costs 32-80 bytes in the gcc build.