win32: fix detection of directories in stat(2)
The implementation of stat(2) detected whether a pathname ending
with a directory separator was valid by checking for the error
code ERROR_INVALID_NAME when GetFileAttributesExA() failed.

This works if the path refers to an actual disk but not if it's on
a share.  In the latter case the glob '*/' incorrectly returned files
that weren't directories.

Add code to handle this case.

Costs 16-32 bytes.

(GitHub issue #381)