lineedit: remove CRs from input
Commit 8ade494aeb (win32: add support for virtual terminal input)
made some changes to how input is handled in the line editing code.

In some circumstances (e.g. 'cat | sh -i') this results in input
being read using fgets(3) rather than from the keyboard.  Remove
CRs from input in this case.

Also, use fgets(3) if stdin is not a tty, but not stdout.