ash: special case in creation of pipeline
When a pipeline is created the parent allocates an array of procstat
structures sufficient for the number of members of the pipeline.
The child processes, however, only see the number of processes
created so far.  Thus, the first child sees that a procstat array
has been created but contains no processes.

This resulted in an invalid pointer to the procstat array in the
first child.  This probably doesn't matter:  if the number of
processes is zero there should be no need to access the array.
Set the pointer to NULL so if it is accessed we'll know about it.