ash: reduce size of forkshell relocation block
The forkshell relocation block had one byte for each byte of the
forkshell structure and funcblock.  It's only necessary to allow
one byte for each pointer.  (One or two bits would be enough, but
would make the code more complex.)

The impact of this change depends on the data and whether the CPU
is 32- or 64-bit.  It typically saves a few Kbytes.

Costs 16-52 bytes.