build system: more clang/llvm tweaks
Linkers associated with clang/llvm may not support the -r option.
This is used to create built-in.o object files.  It turns out that
all such files in busybox-w32 are either empty or only contain one
object file.  The first case is already supported and the second
can be handled by simply copying the object file to built-in.o.
The linker is therefore never invoked with the -r option.

One adjustment is required:  the workaround adopted for GitHub
issue #200 linked the dummy C file with the resource object file.
This is no longer done so only one object file is used.  Since it
was the linking that broke the resource file, copying it is an
equally effective fix for the issue.

Some old linkers don't support the --warn-common option.  The lack
of this option was being detected but it was still sometimes used.