win32: import dirname(3) from mingw-w64
The mingw-w64 project has updated its implementation of dirname(3).
In some circumstances the new version doesn't preserve the type of
the user-supplied top-level directory separator.  As a result of
this the dirname-handles-root test case failed.

Import the new implementation and tweak it to preserve the type of
the separator.

This only affects mingw-w64 versions 12 and above.  Currently only
the aarch64 build using llvm-mingw is affected.