ash: allow nofork applets in Windows on ARM + UCRT
Commit ea8742bc16 (ash: workaround environment issue in Windows
on ARM) prevented nofork applets from being run without a fork
in standalone shell mode in Windows on ARM.  This was due to a
deficiency in the handling of the 'environ' global.

However, the problem is specific to MSVCRT.  If the target is
UCRT nofork applets can be made to work without a fork.