httpd: avoid unnecessary process creation
The httpd server made unnecessary calls to detach from the console.
Avoiding this saves one process creation when a file is being
served and two for a CGI script.

Then again, commit ad4bd5bbd (win32: avoid console windows from CGI
scripts) results in a console host always being created for a CGI
script even if it turns out to be unnecessary.  So the net saving
is only one process for CGI scripts.