build system: avoid full rebuild when EXTRAVERSION changes
The last two commits allow EXTRAVERSION to track the current state
of a git repository.  The build system was unable to determine
which files were affected by changes to EXTRAVERSION and caused
a full rebuild when it changed.

Alter how the version information is passed to the code so only
a handful of files need to be rebuilt when it changes.