tls: use C code for x86_64 sp_256_sub_8_p256_mod()
There's an upstream bug report:

   https://bugs.busybox.net/show_bug.cgi?id=15679

that if BusyBox is built for x86_64 with gcc and the use of OpenSSL
for wget is disabled, failures ensue.  A similar issue also affects
the busybox-w32 build with the comparable configuration.

The problem appears to be in the assembly code for the function
sp_256_sub_8_p256_mod(), as forcing the use of the provided C code
seems to fix both upstream and busybox-w32.  Since my knowledge of
x64_64 assembler is non-existent, someone else will need to determine
the actual issue.  For the moment, just avoid the faulty code.

Saves 32 bytes in the 64-bit build.

(GitHub issue #263)