ash: fix 'read' built-in performance regression
Commits 8e6991733 and b2901ce8e fixed problems with the 'read'
shell built-in when the '-t' option was used.  However, they
result in a performance penalty.  This pipeline:

   seq -w 0 999999 | while read line; do :; done

takes 10 times longer than prior to the changes.

If no timeout is specified don't call poll(2).

Costs 16 bytes in a 32-bit build; 0 in 64-bit.