su: handle restricted disk drivers
Certain tools which allow disk image files and RAM disks to be
mounted were found to lack a feature used to implement realpath(3).
This resulted in a segfault in the 'su' applet when it was run in
such a virtual filesystem.

'su' uses realpath(3) to canonicalise the current directory.  This
is only really required to handle network shares mapped to a drive
letter.  (GitHub issue #148)

- If the call to realpath(3) fails for some reason fall back to
  using the current directory determined by calling getcwd(3).

- If getcwd(3) fails simply don't pass any directory to the shell
  being started by 'su'.

Also, ensure all allocated memory is freed, if required.

(GitHub issue #389)

Adds 16-32 bytes.