win32: improvements to realpath(3)
The previous commit unnecessarily duplicated the path returned
from resolve_symlinks(), resulting in a memory leak.  Thanks to
avih for spotting this.

Even if we can't canonicalise the path in resolve_symlinks() we
can at least return the result of resolving the trailing symlink.
Moreover, this can be done both when the necessary API is missing
and when the filesystem doesn't support it.

Not perfect, but better than nothing, and there's no cost.

This change gets rid of a handful of realpath-related test failures
on ReactOS and Windows XP.

Some background:

Commit 8ebe81483 returned the raw path when the required API was
missing.  This was reverted in commit f902184fa because it caused
an infinite loop (GitHub issue #204).  Commit 31467ddfc fixed the
cause of the loop, which is why it's now safe to reintroduce a
fallback return.

(GitHub commit #389)