make: allow '#' to be escaped with a backslash
POSIX doesn't allow the '#' comment marker to be escaped, though
some implementations do.

As a non-POSIX extension allow '#' to be escaped with a preceding
backslash.

It isn't necessary to escape '#' in macro expansions or command
lines:  these cases are covered by an existing extension.  Commit
0aceca867 (make: comments in macro expansions and command lines)

Adds 16-32 bytes.

(pdpmake GitHub issue 38)