ash: strip path from NOFORK applet names
Commit 26ba73098 (win32: search PATH for missing Unix-style
executables) extended the cases in which PATH would be searched
for executables specified with a Unix-style path.

A side effect of this change was to pass the Unix-style path as
argv[0] to NOFORK applets.  The 'uname' applet examines argv[0]
to determine how to behave.  When invoked as '/bin/uname' it
returned unexpected results.  Other applets may be similarly
affected.

When a NOFORK applet is invoked in evalcommand(), strip any path.

Costs 16 bytes.

(GitHub issue #392)