ash: move setting of current directory
The undocumented '-d' shell option is used to set the current
directory in shells started by the 'su' applet of busybox-w32.
In this case, the shell isn't a login shell.

If a login shell sets the current working directory in /etc/profile
it's possible the user may wish to override this with '-d'.  This
didn't work, though, because the directory is changed before
/etc/profile is processed.

Move the changing of the directory to that specified by '-d' so it
happens after the processing of /etc/profile and ~/.profile.  This
won't affect the intended use of '-d'.

(GitHub issue #403)