make: change how macros are read from the environment
Originally there was no validation of macro names imported from
the environment.  However, process_macros() failed to account for
the additional flag bit at this level.  Thus, for example, SHELL
was imported when it shouldn't have been.

The problem the lack of validation was supposed to address was
that of environment variables with invalid names, which the user
may not have control over, causing a fatal error.

As an alternative, silently ignore variables with invalid names
on import from the environment.