ash: detect console state on every call to options()
Commit 67ed7484be (ash: detect console state on shell start up)
synchronised the noconsole option with the actual state of the
window on shell start up.

This is insufficient.  The user can change the state of the window
independently of the noconsole option, leading to confusion and
unwanted iconification of the window when unrelated 'set' commands
are issued.

Detect the current console state on every call to options().

Saves 16-32 bytes.

(GitHub issue #325)