ash: fix alias expansion followed by '&'
An alias expansion immediately followed by '&' is parsed
incorrectly:

   ~ $ alias x='sleep 2'
   ~ $ x&
   ~ $
   sh: syntax error: unexpected "&"

The sleep happens in the foreground and the '&' is left in the
input buffer.  The same problem occurs in upstream BusyBox but
not dash.

The difference between BusyBox and dash is that BusyBox supports
bash-style output redirection (BASH_REDIR_OUTPUT in the code).
This requires checking for '&>' in readtoken1().

When the end of the alias is found, the '&' and the following
newline are both read to check for '&>'.  Since there's no match
both characters are pushed back.

The alias is then expanded and __pgetc() is called to fetch the
next character.  Since there are none left in the alias string
__pgetc() calls preadbuffer() which pops the string, reverts to
the previous input and recursively calls __pgetc().  This request
is satisified from the pungetc buffer.  But the first __pgetc()
doesn't know this: it sees the character has come from
preadbuffer() so it (incorrectly) updates the pungetc buffer.

Resolve the issue by moving the code to pop the string and fetch
the next character up from preadbuffer() into __pgetc().

Saves 32-48 bytes.

(GitHub issue #413)