make: set $< and $* for target rules
POSIX only requires $< and $* to be set for inference rules, not
target rules.

As an extension allow them to be set for target rules, as in
GNU make.  This may or may not be useful.

In POSIX mode, when $< and $* are only set for inference rules,
they're set to an empty string for target rules.  This avoids
the possibility of stale values being used.

Adds 64-80 bytes.

(GitHub issue #407)