make: add posix_2017 pragma
The pragma 'posix_202x' causes the 202X POSIX standard to be
enforced in POSIX mode.  Add an equivalent 'posix_2017' for the
2017 standard.

There's now a DEFAULT_POSIX_LEVEL preprocessor symbol to configure
the default standard.  This is hardcoded to the 2017 standard but
it can also be set to 202X.

Adds 48 bytes.