make: fix detection of target rules (take 2)
Commit d6b764116 (make: fix detection of target rules) checked
for target rules before macro assignments.  This failed for some
Makefiles generated by autotools because partially defined macros
were expanded while testing for a target rule.

Revert to checking for macro assignments first, but try to detect
if the proposed left hand side of the assignment might form part
of a target rule with an inline command.

Also handle the case where the ';' separator of the inline command
has been obfuscated by putting it in a macro.

Saves 128-160 bytes.

(GitHub pdpmake issues 31, 44)